Examples: Move shared functions into utils module
